move (TouristicContent) change filename to ToursticContentUI
